// src/models/Company.ts
class Company {
    id?: number = 0;
    company_code:                                   string = '';         
    company:                                        string = '';  
    address:                                        string = '';      
    sss_number:                                     string = '';        
    phic_number:                                    string = '';         
    hdmf_number:                                    string = '';         
    tax_number:                                     string = '';       
    current_year_id?:                                number = 0;
    minimum_overtime_hours?:                         number = 0;
    create_by:                                      number = 0;
    created_date:                                   Date   = new Date ();
    update_by:                                      number = 0;
    update_date:                                    Date   = new Date ();
    is_locked?:                                      boolean = false;
    funding_account:                                string  = '';        
    sss_account_id:                                 number = 0;
    hdmf_account_id:                                number = 0;
    phic_account_id:                                number = 0;
    tax_account_id:                                 number = 0;
    image_url:                                      string = '';         
    number_of_branches?:                             number = 0;
    Requirements?:                                   string = '';         
    is_special_holiday_without_logs:                boolean = false;
    is_overtime_taxable:                           boolean = false;
    is_restday_taxable:                             boolean = false;
    is_night_diff_taxable:                          boolean = false;
    is_premium_taxable?:                             boolean = false;
    is_holiday_taxable:                             boolean = false;
    is_additional_allowance_taxable:                boolean = false;
    is_shown_non_taxable_income:                    boolean = false;
    is_show_loan_balance:                           boolean = false;
    is_shown_monthly_rate:                          boolean = false;
    is_payroll_type_fixed_compute_as_variable:      boolean = false;
    is_show_approved_by:                            boolean = false;
    is_show_daily_rate:                             boolean = false;
    is_regular_holiday_included_sss_contribution:   boolean = false;
    is_overtime_convertible_to_leave_credits?:       boolean = false;
    is_show_payslip_other_income_detail:            boolean = false;
    is_phic_computation_based_on_payroll_rate:      boolean = false;
    is_thirteen_month_included_restday_pay:         boolean = false;
    is_thirteen_month_included_overtime_pay:        boolean = false;
    is_thirteen_month_included_special_holiday_pay: boolean = false;
    is_thirteen_month_included_regular_holiday_pay: boolean = false;
    is_thirteen_month_included_absent_deduction:    boolean = false;
    is_thirteen_month_included_late_deduction:      boolean = false;
    is_thirteen_month_included_undertime_deduction: boolean = false;
    is_payslip_daily_rate_from_payrolline?:          boolean = false;
    is_shown_holiday_amount_payslip?:                boolean = false;
    is_multiple_breaks?:                             boolean = false;
}
export default Company;